pyproject.toml
[project]
name = "python-docker-compose-example"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "psycopg2-binary>=2.9.10",
]
